[EGD-8079] CPU frequency conversion PWM fix
Fixed CPU frequency conversion.

+++ module-bsp/board/rt1051/bsp/eink_frontlight/eink_frontlight.cpp
@@ -65,7 +65,7 @@ namespace bsp::eink_frontlight
 
     void updateClockFrequency(CpuFrequencyMHz newFrequency)
     {
-        pwm->UpdateClockFrequency(static_cast<std::uint32_t>(newFrequency));
+        pwm->UpdateClockFrequency(newFrequency);
     }
 
 } // namespace bsp::eink_frontlight
